Add files via upload
Nothing produced in this file goes on to be used in modelling - here we just explored the Foursquare API, how to use it, the JSON it provides and all its columns. including this provides insight as to the entire process. 

'bike_poi_radius_fsq.ipynb' is the foursquare data file that will be used later in analysis.